move update event decryption methods to extension
