+++ Sources/Helpers/EncryptionSessionDirectory+UpdateEvents.swift
@@ -108,9 +108,12 @@ extension EncryptionSessionsDirectory {
     /// Returns the decrypted version of an update event. This is generated by decrypting the encrypted version
     /// and creating a new event with the decrypted data in the expected payload keys
     fileprivate func decryptedUpdateEvent(for event: ZMUpdateEvent, sender: UserClient) throws -> (createdNewSession: Bool, event: ZMUpdateEvent)? {
-        guard let result = try self.decryptedData(event, client: sender),
+        guard
+            let result = try self.decryptedData(event, client: sender),
             let decryptedEvent = event.decryptedEvent(decryptedData: result.decryptedData)
-        else { return nil }
+        else {
+            return nil
+        }
         return (createdNewSession: result.createdNewSession, event: decryptedEvent)
     }
 
@@ -196,58 +199,5 @@ extension ZMUpdateEvent {
 
     fileprivate var externalStringCount: Int {
         return (eventData?["data"] as? String)?.count ?? 0
-    }
-
-    /// Returns a decrypted version of self, injecting the decrypted data
-    /// in its payload and wrapping the payload in a new updateEvent
-    private func decryptedEvent(decryptedData: Data) -> ZMUpdateEvent? {
-        guard
-            var payload = self.payload as? [String: Any],
-            var eventData = payload["data"] as? [String: Any]
-        else {
-            return nil
-        }
-
-        if self.type == .conversationOtrMessageAdd, let wrappedData = eventData["data"] as? String {
-            payload["external"] = wrappedData
-        }
-
-        eventData[self.plaintextPayloadKey] = decryptedData.base64String()
-        payload["data"] = eventData
-
-        return decryptedEvent(payload: payload)
-    }
-
-    // TODO: Move this somewhere else
-    func decryptedMLSEvent(decryptedData: Data) -> ZMUpdateEvent? {
-        guard var payload = self.payload as? [String: Any] else {
-            return nil
-        }
-
-        payload["data"] = ["text": decryptedData.base64EncodedString()]
-
-        return decryptedEvent(payload: payload)
-    }
-
-    func decryptedEvent(payload: [String: Any]) -> ZMUpdateEvent? {
-        let decryptedEvent = ZMUpdateEvent.decryptedUpdateEvent(fromEventStreamPayload: payload as NSDictionary, uuid: uuid, transient: false, source: source)
-
-        if !self.debugInformation.isEmpty {
-            decryptedEvent?.appendDebugInformation(debugInformation)
-        }
-
-        return decryptedEvent
-    }
-
-    /// Payload dictionary key that holds the plaintext (protobuf) data
-    fileprivate var plaintextPayloadKey: String {
-        switch self.type {
-        case .conversationOtrMessageAdd:
-            return "text"
-        case .conversationOtrAssetAdd:
-            return "info"
-        default:
-            fatal("Decrypting wrong type of event")
-        }
     }
 }
